requirements.txt
telethon==1.31.1
python-dotenv==1.0.0